renaming of class, add annotations to join columns from 2 tables
